stranger v0.1.0 Stranger.Lobby

Holds the IDs and Channel PIDs of all strangers in the lobby channel.

Monitors the Channel PID so that if the stranger leaves the channel they’re removed here.

Functions

find(id)

Find a Channel PID of a stranger with the id.

handle_call(msg, from, state)

Removes both id and pid references from state.

handle_info(arg1, state)

Removes both id and pid references from state if the linked process goes down.

join(id, pid)

Adds a stranger with the id and pid.

leave(id)

Removes a stranger with the id.

start_link()

strangers()

Retrieve the id of all connected strangers.