stranger v0.1.0 Stranger.Notification

Broadcasts notifications when anyone connects/disconnects from the Notification channel.

Functions

count()

Returns the number of strangers in the channel (pids).

handle_info(arg1, state)

Called when someone leaves the notification channel. Removes them from the list of pids and broadcasts a notification message.

monitor(pid)

Monitors the pid and broadcasts a notification message.

start_link()