stranger v0.0.8 Stranger.Room

Each room is for a chat between two strangers.

Only strangers within the ids list are allowed to join.

Functions

create(name, ids)

Helper method for calling Stranger.Room.Supervisor.start_child/2.

handle_call(msg, from, state)

The ID is only allowed to join if the ID is present in the ids list.

join(name, id, pid)

Attempts to the join the room. If successful will return {:ok, pid}.

Failing to join will return {:error, reason}.

start_link(name, ids)

Starts the GenServer, named globally based on the room name.