stranger v0.0.8 Stranger.Router.Helpers

Module with named helpers generated from Stranger.Router.

Functions

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.